+++ app/src/main/java/com/immobylette/appmobile/data/enum/ElementState.kt
@@ -1,0 +1,10 @@
+package com.immobylette.appmobile.data.enum
+
+enum class ElementState(val label: String = "") {
+    NEW("Nouveau"),
+    VERY_GOOD("Très bon"),
+    GOOD("Bon"),
+    SO_SO("Moyen"),
+    BAD("Mauvais"),
+    VERY_BAD("Très mauvais")
+}
