+++ app/src/main/java/com/immobylette/appmobile/ui/shared/component/Tip.kt
@@ -3,7 +3,7 @@ package com.immobylette.appmobile.ui.shared.component
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
@@ -15,7 +15,6 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.times
 import com.immobylette.appmobile.ui.shared.theme.ImmobyletteappmobileTheme
@@ -24,31 +23,19 @@ import com.immobylette.appmobile.ui.shared.theme.ImmobyletteappmobileTheme
 fun Tip(
     text: String,
     modifier: Modifier = Modifier,
-    style: TextStyle = MaterialTheme.typography.headlineLarge,
+    style: TextStyle = MaterialTheme.typography.headlineMedium
 ) {
-    val padding: Dp
-    val width: Dp
-
-    if (style != MaterialTheme.typography.labelSmall){
-        padding = 10.dp
-        width = (text.length * 15.dp) + 10.dp
-    }else {
-        padding = 5.dp
-        width = (text.length * 10.dp) + 5.dp
-    }
-
     Box(
         modifier = modifier
             .clip(RoundedCornerShape(30.dp))
             .background(MaterialTheme.colorScheme.primary)
-            .padding(padding)
-            .width(width),
+            .width((text.length * 15.dp) + 10.dp),
         contentAlignment = Alignment.Center
     ) {
         Text(
             text = text,
-            style = style,
-            color = Color.White
+            color = Color.White,
+            style = style
         )
     }
 }
@@ -60,10 +47,10 @@ fun TipPreview(){
     ImmobyletteappmobileTheme {
         Row {
             Tip("1/4")
-            Tip("1", style = MaterialTheme.typography.headlineSmall)
+            Tip("1", modifier = Modifier.size(20.dp), style = MaterialTheme.typography.labelSmall)
             Tip("20/20")
-            Tip("1", style = MaterialTheme.typography.labelSmall)
-            Tip("10", style = MaterialTheme.typography.labelSmall)
+            Tip("1")
+            Tip("10")
         }
     }
 }
